# Avoid overriding CipherSpi::engineDoFinal(ByteBuffer, ByteBuffer)